Add more clang-tidy checks
* Enables bugprone- checks.
* Enables a couple more modernize checks.

Most of the changes are avoiding implicit narrowing conversions.
There was a bug caught and fixed fixed with the string comparison in iso_fs.

bugprone-suspicious-enum-usage was disabled due to a large number of
false positives with wxWidgets.

+++ CorsixTH/Src/sdl_core.cpp
@@ -70,8 +70,8 @@ class fps_ctrl {
   bool limit_fps;
   bool track_fps;
 
-  int q_front;
-  int q_back;
+  size_t q_front;
+  size_t q_back;
   int frame_count;
   std::array<Uint32, 4096> frame_time;
