+++ CorsixTH/Src/xmi2mid.cpp
@@ -66,7 +66,7 @@ class memory_buffer {
     return true;
   }
 
-  bool skip(int distance) {
+  bool skip(ssize_t distance) {
     if (distance < 0) {
       if (pointer + distance < data) return false;
     }
@@ -95,7 +95,7 @@ class memory_buffer {
     return true;
   }
 
-  unsigned int read_big_endian_uint24() {
+  uint32_t read_big_endian_uint24() {
     uint8_t iByte0, iByte1, iByte2;
     if (read(iByte0) && read(iByte1) && read(iByte2))
       return (((iByte0 << 8) | iByte1) << 8) | iByte2;
@@ -103,7 +103,7 @@ class memory_buffer {
       return 0;
   }
 
-  unsigned int read_variable_length_uint() {
+  uint32_t read_variable_length_uint() {
     unsigned int iValue = 0;
     uint8_t iByte;
     for (int i = 0; i < 4; ++i) {
@@ -121,7 +121,7 @@ class memory_buffer {
 
   template <class T>
   bool write(const T* values, size_t count) {
-    if (!skip(static_cast<int>(sizeof(T) * count))) return false;
+    if (!skip(static_cast<ssize_t>(sizeof(T) * count))) return false;
     std::memcpy(pointer - sizeof(T) * count, values, sizeof(T) * count);
     return true;
   }
@@ -219,7 +219,7 @@ uint8_t* transcode_xmi_to_midi(const unsigned char* xmi_data, size_t xmi_length,
   midi_token_list lstTokens;
   midi_token* pToken;
   int iTokenTime = 0;
-  int iTempo = 500000;
+  uint32_t iTempo = 500000;
   bool bTempoSet = false;
   bool bEnd = false;
   uint8_t iTokenType, iExtendedType;
@@ -253,7 +253,9 @@ uint8_t* transcode_xmi_to_midi(const unsigned char* xmi_data, size_t xmi_length,
         pToken->data = iExtendedType;
         if (!bufInput.skip(1)) return nullptr;
         pToken = lstTokens.append(
-            iTokenTime + bufInput.read_variable_length_uint() * 3, iTokenType);
+            iTokenTime +
+                static_cast<int>(bufInput.read_variable_length_uint()) * 3,
+            iTokenType);
         pToken->data = iExtendedType;
         pToken->buffer = "\0";
         break;
@@ -272,7 +274,8 @@ uint8_t* transcode_xmi_to_midi(const unsigned char* xmi_data, size_t xmi_length,
               bufInput.skip(-4);
             } else {
               lstTokens.pop_back();
-              if (!bufInput.skip(bufInput.read_variable_length_uint()))
+              if (!bufInput.skip(static_cast<ssize_t>(
+                      bufInput.read_variable_length_uint())))
                 return nullptr;
               break;
             }
